+++ test/concrete/ob/OrderBook.takeOrder.sameToken.t.sol
@@ -38,8 +38,8 @@ contract OrderBookTakeOrderSameTokenTest is OrderBookExternalRealTest {
 
         TakeOrdersConfigV4 memory takeOrdersConfig = TakeOrdersConfigV4({
             minimumInput: LibDecimalFloat.packLossless(0, 0),
-            maximumInput: LibDecimalFloat.packLossless(type(int256).max, 0),
-            maximumIORatio: LibDecimalFloat.packLossless(type(int256).max, 0),
+            maximumInput: LibDecimalFloat.packLossless(type(int224).max, 0),
+            maximumIORatio: LibDecimalFloat.packLossless(type(int224).max, 0),
             orders: takeOrders,
             data: ""
         });
