+++ test/concrete/ob/OrderBook.takeOrder.tokenMismatch.t.sol
@@ -42,8 +42,8 @@ contract OrderBookTakeOrderTokenMismatchTest is OrderBookExternalRealTest {
         aOutputIOIndex = bound(aOutputIOIndex, 0, a.validOutputs.length - 1);
         vm.assume(b.validOutputs.length > 0);
         bOutputIOIndex = bound(bOutputIOIndex, 0, b.validOutputs.length - 1);
-        maxTakerInput18 = bound(maxTakerInput18, 1, uint256(type(int256).max));
-        maxIORatio = bound(maxIORatio, 0, uint256(type(int256).max));
+        maxTakerInput18 = bound(maxTakerInput18, 1, uint256(int256(type(int224).max)));
+        maxIORatio = bound(maxIORatio, 0, uint256(int256(type(int224).max)));
 
         vm.assume(a.validInputs[aInputIOIndex].token != a.validOutputs[aOutputIOIndex].token);
 
@@ -87,8 +87,8 @@ contract OrderBookTakeOrderTokenMismatchTest is OrderBookExternalRealTest {
         aOutputIOIndex = bound(aOutputIOIndex, 0, a.validOutputs.length - 1);
         vm.assume(b.validOutputs.length > 0);
         bOutputIOIndex = bound(bOutputIOIndex, 0, b.validOutputs.length - 1);
-        maxTakerInput18 = bound(maxTakerInput18, 1, uint256(type(int256).max));
-        maxIORatio18 = bound(maxIORatio18, 0, uint256(type(int256).max));
+        maxTakerInput18 = bound(maxTakerInput18, 1, uint256(int256(type(int224).max)));
+        maxIORatio18 = bound(maxIORatio18, 0, uint256(int256(type(int224).max)));
 
         vm.assume(a.validOutputs[aOutputIOIndex].token != a.validInputs[aInputIOIndex].token);
